refactor(modules): consolidate the workspace subsystem into src/modules/workspace/
Relocate the workspace subsystem — workspace.c plus its provider, runner,
scope, turn, mirror, manifest, and handle units, and cli_workspace_serve —
from src/ and src/server/ into src/modules/workspace/, moving the 11
matching headers out of src/headers/ alongside them. This mirrors the
economizer/memory/git module layout (module owns its .c and .h at the
module root).

Platform provider implementations (posix/ and windows/ workspace_provider.c)
stay in their platform trees since they are two same-named files that
cannot share a flat module dir; worktree_gc remains a separate subsystem.

Add -Imodules/workspace to the compiler include path and re-prefix every
source and object reference across src/Makefile, src/tests/Rules.mk, and
CMakeLists.txt (leaving the posix/windows platform providers in place).
The three path-qualified headers/workspace_manifest.h includes become bare
so they resolve via the new include dir.
