+++ src/cmd_infra.c
@@ -16,7 +16,7 @@
 #include "cJSON.h"
 #include "mcp_git.h"
 #include "git_verify.h"
-#include "headers/workspace_manifest.h"
+#include "workspace_manifest.h"
 #include <unistd.h>
 #include <sys/stat.h>
 #include <ctype.h>
