+++ src/modules/workspace/workspace_manifest.h
