+++ src/modules/workspace/workspace_mirror.c
