+++ src/modules/workspace/workspace_provider_detached.c
