+++ src/modules/workspace/workspace_provider_detached.h
