+++ src/modules/workspace/workspace_scope.c
