+++ src/modules/workspace/workspace_scope.h
